morning update from mack

+++ app/src/main/java/com/mykaimeal/planner/fragment/mainfragment/profilesetting/orderhistoryscreen/model/OrderHistoryModel.kt
@@ -35,7 +35,7 @@ data class FinalQuote(
     val store_address: String?,
     val store_id: String?,
     val tip: Int,
-    val total_with_tip: Int
+    val total_with_tip: Double
 )
 
 data class AddedFees(
